telegram bot/requirements.txt
python-telegram-bot==13.4.1
pandas==1.3.4